backend/apps/reports/views.py
from rest_framework import generics, permissions
from .models import Report
from rest_framework import serializers


class ReportSerializer(serializers.ModelSerializer):
    class Meta:
        model = Report
        fields = '__all__'
        read_only_fields = ['id', 'organization', 'created_at', 'updated_at']


class ReportListView(generics.ListAPIView):
    serializer_class = ReportSerializer
    permission_classes = [permissions.IsAuthenticated]

    def get_queryset(self):
        return Report.objects.filter(
            organization=self.request.user.organization,
            is_deleted=False
        )