backend/requirements.txt
Django==4.2.15
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.10.0
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
dj-database-url==2.1.0
whitenoise==6.5.0
gunicorn==21.2.0
python-decouple==3.8
Pillow==10.0.1
factory-boy==3.3.0
pytest==7.4.2
pytest-django==4.6.0
pytest-cov==4.1.0
flake8==6.1.0
black==23.9.1
isort==5.12.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.16.1